+++ examples/ai-functions/src/generate-speech/fal-voice.ts
@@ -1,4 +1,4 @@
-import { fal } from '@ai-sdk/fal';
+import { fal, type FalSpeechModelOptions } from '@ai-sdk/fal';
 import { experimental_generateSpeech as generateSpeech } from 'ai';
 import { saveAudioFile } from '../lib/save-audio';
 import { run } from '../lib/run';
@@ -13,7 +13,7 @@ run(async () => {
           'Bubbly and excitable female pop star interviewee, youthful, slightly breathless, and very enthusiastic',
         preview_text:
           "Oh my gosh, hi. It's like so amazing to be here. This new endpoint just dropped on fal and the results have been like totally incredible. Use it now, It's gonna be like epic!",
-      },
+      } satisfies FalSpeechModelOptions,
     },
   });